[Refactor] PlaceListFragment의 RecyclerView Adapter 클래스 하나로 통일
