+++ app/src/main/java/com/twoday/todaytrip/place_list_adapter/PlaceListRecyclerViewAdapter.kt
@@ -1,7 +1,9 @@
 package com.twoday.todaytrip.place_list_adapter
 
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.core.view.isVisible
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
@@ -11,9 +13,9 @@ import com.twoday.todaytrip.databinding.ItemPlaceListBinding
 import com.twoday.todaytrip.tourData.TourItem
 import com.twoday.todaytrip.utils.ContentIdPrefUtil
 
-class EventRecyclerViewAdapter :
-    ListAdapter<TourItem, EventRecyclerViewAdapter.Holder>(TourItemDiffCallback) {
-    private val TAG = "EventRecyclerViewAdapter"
+class PlaceListRecyclerViewAdapter :
+    ListAdapter<TourItem, PlaceListRecyclerViewAdapter.Holder>(TourItemDiffCallback) {
+    private val TAG = "PlaceListRecyclerViewAdapter"
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
         val binding =
@@ -23,7 +25,7 @@ class EventRecyclerViewAdapter :
 
     override fun onBindViewHolder(holder: Holder, position: Int) {
         holder.run {
-            //initOnClickListener(getItem(position))
+            initOnClickListener(getItem(position))
             bind(getItem(position))
         }
     }
@@ -40,8 +42,19 @@ class EventRecyclerViewAdapter :
         private val titleTextView = binding.tvItemPlaceListTitle
         private val addressTextView = binding.tvItemPlaceListAddress
 
-        //TODO 영업시간, 휴무일 표시하기
-        //private val addButton = binding.btnItemPlaceListAdd
+        private val timeLabelTextViewList = listOf(
+            binding.tvItemPlaceListTimeLabel1,
+            binding.tvItemPlaceListTimeLabel2,
+            binding.tvItemPlaceListTimeLabel3
+        )
+        private val timeTextViewList = listOf(
+            binding.tvItemPlaceListTime1,
+            binding.tvItemPlaceListTime2,
+            binding.tvItemPlaceListTime3
+        )
+
+        private val addLayout = binding.layoutItemPlaceListAdd
+        private val addTextView = binding.tvItemPlaceListAdd
 
         fun bind(item: TourItem) {
             item.getThumbnailImage()?.let { url ->
@@ -51,36 +64,47 @@ class EventRecyclerViewAdapter :
                     .into(firstImageView)
             }
             firstImageView.clipToOutline = true
+
             titleTextView.text = item.getTitle()
             addressTextView.text = item.getAddress()
 
-            //setAddButtonUI(item.isAdded)
+            item.getTimeInfoWithLabel().forEachIndexed { index, pair ->
+                Log.d(TAG, "getTimeInfoWithLabel) index: $index, pair: $pair")
+                with(timeLabelTextViewList[index]) {
+                    isVisible = true
+                    text = pair.first
+                }
+                with(timeTextViewList[index]) {
+                    isVisible = true
+                    text = pair.second
+                }
+            }
+
+            setAddButtonUI(item.isAdded)
         }
-/*
+
         fun initOnClickListener(item: TourItem) {
-            this.addButton.setOnClickListener {
+            this.addLayout.setOnClickListener {
                 OnTourItemClickListener.onTourItemClick(item)
                 setAddButtonUI(item.isAdded)
             }
         }
 
         private fun setAddButtonUI(isAdded: Boolean) {
-            addButton.background = MyApplication.appContext!!.resources.getDrawable(
+            addLayout.background = MyApplication.appContext!!.resources.getDrawable(
                 if (isAdded) R.drawable.shape_white_with_border_radius_10
                 else R.drawable.shape_mainblue_10_radius
             )
-            addButton.text = MyApplication.appContext!!.resources.getText(
+            addTextView.text = MyApplication.appContext!!.resources.getText(
                 if (isAdded) R.string.item_place_list_remove
                 else R.string.item_place_list_add
             )
-            addButton.setTextColor(
+            addTextView.setTextColor(
                 MyApplication.appContext!!.resources.getColor(
                     if (isAdded) R.color.main_blue
                     else R.color.white
                 )
             )
         }
-
- */
     }
 }
